[et] Move XcodeGen types to separate file
