+++ tools/expotools/src/prebuilds/XcodeProject.ts
@@ -3,9 +3,10 @@ import path from 'path';
 import fs from 'fs-extra';
 
 import { spawnAsync } from '../Utils';
-import { generateXcodeProjectAsync, ProjectSpec } from './XcodeGen';
+import { generateXcodeProjectAsync } from './XcodeGen';
 import { Flavor, Framework, XcodebuildSettings } from './XcodeProject.types';
 import { formatXcodeBuildOutput } from '../Formatter';
+import { ProjectSpec } from './XcodeGen.types';
 
 /**
  * Path to the shared derived data directory.
